chore: Update homepage URL and add predeploy and deploy scripts
This commit updates the homepage URL in the package.json file to the correct GitHub Pages URL. It also adds predeploy and deploy scripts to automate the build and deployment process using the gh-pages package. These changes ensure that the website is deployed correctly and can be accessed by users. Deploy app